+++ CHANGELOG.md
@@ -3,10 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.0.10] - 2015-10-15
+### Added
+- Checks table. Parsing is still buggy though.
+
 ### Changed
 - Reworked the way the program generates default values.
 - Fixed (MAX) bug for user-defined types.
 - Fixed ASC or DESC inclusion in indeces where required.
+- Fixed sequence data type and maximum value.
+- Put table names in double quotation marks in case the name is not PG kosher. 
 
 ## [0.0.9] - 2015-10-12
 ### Added